Begin rework
