+++ src/divio_docs_gen/__init__.py
@@ -1,4 +1,4 @@
 # Import for use in pyproject.toml[project.scripts]
-from .index import generate_docs
+from .index import main#generate_docs, parse_all_sections_markdown_content
 
 """Implements standalone divio_docs_gen usage"""
