+++ src/divio_docs_gen/__main__.py
@@ -1,6 +1,6 @@
-from .index import generate_docs
+from .index import main
 
 """Implements python3 -m divio_docs_gen usage"""
 
 if __name__ == "__main__":
-    generate_docs()
+    main()
